# Faster reverse for tuples