--- a/base/tuple.jl
+++ b/base/tuple.jl
@@ -172,7 +172,13 @@ end
 isempty(x::Tuple{}) = true
 isempty(x::Tuple) = false
 
-reverse(t::Tuple) = tuple(reverse(collect(t))...)
+reverse(t::Tuple{}) = t
+reverse(t::Tuple{T}) where {T} = t
+reverse(t::Tuple{T,U}) where {T,U} = t[2], t[1]
+reverse(t::Tuple{T,U,V}) where {T,U,V} = t[3], t[2], t[1]
+reverse(t::Tuple{T,U,V,W}) where {T,U,V,W} = t[4], t[3], t[2], t[1]
+reverse(t::NTuple{N}) where N = ([t[i] for i in length(t):-1:1]...)::NTuple{N}
+reverse(t::Tuple{Any,Any,Any,Any,Any,Vararg{Any}}) = ([t[i] for i in length(t):-1:1]...)
 
 ## specialized reduction ##