The proposed change in `reverse(t::Tuple)` avoids recursion and in particular calling `revargs` for varying number of arguments which especially slows down the first execution of calls like `@time reverse(tuple(1:1000...));`. The change makes definition of `revargs` unnecessary.