@TotalVerb 2016-05-26
:+1:. The original code was a cute Scheme analogy, but I don't think it's right for Julia.

My benchmarking indicates that

```julia
reverse3(t::Tuple) = ([t[i] for i in length(t):-1:1]...)
```
is _slightly_ faster. It's a bit messier though.