@yuyichao 2016-05-26
https://github.com/JuliaLang/julia/issues/15695 note that this is not type stable and will be very bad for small tuples