@JeffBezanson 2016-05-26
Yes, unfortunately this definition will be a disaster for small tuples.