@TotalVerb 2016-05-26
For (heterogenous and homogenous) tuples of length 5 or greater, the current `reverse` is worse than the modified proposed one. You're right that it's better for small tuples.