@TotalVerb 2016-05-27
This seems to work:

```julia
reverse(t::Tuple{}) = t
reverse{T}(t::Tuple{T}) = t
reverse{T,U}(t::Tuple{T,U}) = t[2], t[1]
reverse{T,U,V}(t::Tuple{T,U,V}) = t[3], t[2], t[1]
reverse{T,U,V,W}(t::Tuple{T,U,V,W}) = t[4], t[3], t[2], t[1]
reverse(t::Tuple) = ([t[i] for i in length(t):-1:1]...)
```

Is this considered too messy?
rev (generic function with 6 methods)