@JeffBezanson 2016-05-27
Yes, we will probably need to go with something like that. A slightly better trick can be seen in https://github.com/JuliaLang/julia/pull/16460/files#diff-fd99c9a15dadc23d9200069dadafb8bbR90
It's messy but I don't know a better way to do it so far.