@TotalVerb 2016-05-27
That is indeed a nice trick. So only one additional line is needed:

```julia
reverse(t::Tuple{Any,Any,Any,Any,Any,Vararg{Any}}) = ([t[i] for i in length(t):-1:1]...)
```

This seems to have the best performance. I wonder why the size of tuple to reach performance parity is so much smaller in this case than with `map`.