@Jutho 2016-05-27
It seems that the definition:
```julia
reverse{N}(t::NTuple{N,Any}) = tuple(reverse(collect(t))...)::NTuple{N,Any}
```
is at least type-stable for homogeneous tuples. It infers the type from the `reverse(collect(t))` construction and combines it with the length in the type assertion.