@bkamins 2016-05-28
Thank you all for dedication to squeeze maximum performance out of every part of Julia :+1:.

Having learned again how little I know about Julia I have filed an issue #16631, and not PR :), for two other cases that have performance issues related to large tuples.