@musm 2016-12-04
the proposals in this thread seem to be an improvement over the original, should someone update this PR?