@bkamins 2016-12-04
👍. I did not update PR myself as I feel that someone with better understanding of Julia should choose the best approach to fix `reverse`.