@JeffBezanson 2016-12-06
The idea here looks good to me, since we just have to add one line: https://github.com/JuliaLang/julia/pull/16604#issuecomment-222062170