@TotalVerb 2016-12-06
@JeffBezanson There is still the type stability problem to deal with though, which I didn't measure in my benchmarks (since it only affects downstream consumers) and is harder to measure in general. Perhaps we should keep the same cutoff as `map`?