@TotalVerb 2016-12-06
There is also @Jutho's suggestion which preserves type stability for homogenous tuples

```julia
reverse{N}(t::NTuple{N,Any}) = tuple(reverse(collect(t))...)::NTuple{N,Any}
```

which may perhaps be useful to work in.