@JeffBezanson 2016-12-06
I think it will work to add `::NTuple{nfields(t),Any}` to the comprehension version.