@martinholters 2016-12-07
How about
```Julia
@generated reverse(x::Tuple) = Expr(:tuple, (:(x[$i]) for i in nfields(x):-1:1)...)
```
? Fast and type-stable.