@TotalVerb 2016-12-07
This issue is mostly about compile time, and it's highly doubtful that has better compile time.