@martinholters 2016-12-07
> This issue is mostly about compile time, and it's highly doubtful that has better compile time.

Oh. Really? With the OP's example:
```Julia
julia> @benchmark reverse($(tuple(1:1000...))) # master
BenchmarkTools.Trial: 
  memory estimate:  18.17 mb
  allocs estimate:  373638
  --------------
  minimum time:     18.886 ms (0.00% GC)
  median time:      20.347 ms (0.00% GC)
  mean time:        23.854 ms (14.99% GC)
  maximum time:     105.331 ms (81.43% GC)
  --------------
  samples:          210
  evals/sample:     1
  time tolerance:   5.00%
  memory tolerance: 1.00%

julia> @benchmark myreverse($(tuple(1:1000...))) # as proposed above
BenchmarkTools.Trial: 
  memory estimate:  0.00 bytes
  allocs estimate:  0
  --------------
  minimum time:     709.319 ns (0.00% GC)
  median time:      714.773 ns (0.00% GC)
  mean time:        727.437 ns (0.00% GC)
  maximum time:     1.767 μs (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     141
  time tolerance:   5.00%
  memory tolerance: 1.00%
```
Of course, this does not include compile time, but the execution time is improved dramatically.