@KristofferC 2016-12-07
This reminds me of: https://github.com/JuliaLang/julia/pull/15702#issuecomment-203896961. 

Here is another benchmark calling the reverse function for the first time (so runtime + compilation time):

```
v = Float64[]
for i in 1:100
       t = (rand(i)...)    
       push!(v, @elapsed reverse(t))
end
```


![image](https://cloud.githubusercontent.com/assets/1282691/20960807/701d5fd4-bc62-11e6-8149-da42c87b4bfc.png)