@bkamins 2017-09-13
I wanted to go back to this PR and resolve it.
Given the above discussion my recommendation would be:
* I opt against `@generated` version as it generates StackOverflow for very large tuples (and is unacceptably slow to compile for for large tuples below StackOverflow level)
* Have specialized methods for small tuples (I chose up to 4 elements - as in recommendation above) which is type stable
* Have a separate method for `NTuple` as it is type stable and the most probable use case for large tuples will be `NTuple`
* Accept that for large tuples `reverse` might not be type stable

As an additional issue I would like to ask what is the recommended place to document `reverse` in general, as now `help(reverse)` prints help only for `AbstractString`.