@TotalVerb 2017-09-15
As @KristofferC pointed out, the cutoff here should probably be something bigger than 18 and not 4.