@bkamins 2017-09-15
@TotalVerb I agree that the break-point is at ~20 for single execution of `reverse`, and even for 100 elements in a tuple the generation of specialized function is not prohibitively slow (which would give a benefit when `reverse` is called many times).
I thought that as we have to settle for a suboptimal solution then it is best to keep it relatively simple.

Of course it is simple to extend the definitions to generation of a larger number predefined reverses - would you judge it worthwhile? (and the question is if having very many such definitions would not have a negative impact on Julia method dispatch code speed in general - here I do not know)