@TotalVerb 2017-09-15
@bkamins As Jeff mentioned, we may be able to keep the existing definition and simply add a new one, like

```julia
reverse(t::Tuple{Any,Any,Any,Any,Any,Any,Any,Any,Any,Any,Any,Any,Any,Any,Any,Any,Any,Any,Vararg{Any}}) = ([t[i] for i in length(t):-1:1]...)::NTuple{nfields(t), Any}
```

to cover long cases. However this doesn't preserve the type stability of homogenous tuples.