@bkamins 2017-09-16
@TotalVerb I have now unrolled the definitions for tuples not greater than 20. For greater ones the code for homogeneous tuples is type stable. Is this is what you had in mind?