@TotalVerb 2017-09-16
Looks much better. @JeffBezanson and others more familiar with this should probably have a look.