@bkamins 2017-09-24
Small comment - you need `NTuple{nfields(t), Any}` (as you proposed above) for it to work.

The reason is performance. This is a test I use (`new_reverse` is my proposed implementation):
```
function test(i)
    println("Testing tuple size ", i+1)
    for (e, s) in [("a", "mixed"), (0, "homogeneous")]
        t = (e, 1:i...)
        print("\tOld $s:")
        @btime reverse($t)
        print("\tNew $s:")
        @btime new_reverse($t)
    end
end

foreach(test, 1:3:25)
```
which produces:
```
Testing tuple size 2
        Old mixed:  10.264 ns (1 allocation: 32 bytes)
        New mixed:  10.263 ns (1 allocation: 32 bytes)
        Old homogeneous:  1.866 ns (0 allocations: 0 bytes)
        New homogeneous:  1.866 ns (0 allocations: 0 bytes)
Testing tuple size 5
        Old mixed:  1.819 μs (9 allocations: 240 bytes)
        New mixed:  12.596 ns (1 allocation: 48 bytes)
        Old homogeneous:  1.773 μs (9 allocations: 240 bytes)
        New homogeneous:  2.332 ns (0 allocations: 0 bytes)
Testing tuple size 8
        Old mixed:  2.955 μs (14 allocations: 480 bytes)
        New mixed:  18.681 ns (1 allocation: 80 bytes)
        Old homogeneous:  2.903 μs (13 allocations: 400 bytes)
        New homogeneous:  4.198 ns (0 allocations: 0 bytes)
Testing tuple size 11
        Old mixed:  5.443 μs (21 allocations: 816 bytes)
        New mixed:  21.525 ns (1 allocation: 96 bytes)
        Old homogeneous:  5.521 μs (21 allocations: 816 bytes)
        New homogeneous:  6.531 ns (0 allocations: 0 bytes)
Testing tuple size 14
        Old mixed:  6.881 μs (26 allocations: 1.17 KiB)
        New mixed:  22.860 ns (1 allocation: 128 bytes)
        Old homogeneous:  6.881 μs (25 allocations: 1.05 KiB)
        New homogeneous:  8.864 ns (0 allocations: 0 bytes)
Testing tuple size 17
        Old mixed:  3.791 μs (13 allocations: 1.08 KiB)
        New mixed:  22.393 ns (1 allocation: 144 bytes)
        Old homogeneous:  1.120 μs (2 allocations: 368 bytes)
        New homogeneous:  11.196 ns (0 allocations: 0 bytes)
Testing tuple size 20
        Old mixed:  4.332 μs (13 allocations: 1.14 KiB)
        New mixed:  27.992 ns (1 allocation: 176 bytes)
        Old homogeneous:  1.260 μs (2 allocations: 416 bytes)
        New homogeneous:  13.529 ns (0 allocations: 0 bytes)
Testing tuple size 23
        Old mixed:  4.199 μs (13 allocations: 1.34 KiB)
        New mixed:  3.674 μs (13 allocations: 1.34 KiB)
        Old homogeneous:  1.399 μs (2 allocations: 528 bytes)
        New homogeneous:  884.458 ns (2 allocations: 528 bytes)
Testing tuple size 26
        Old mixed:  4.332 μs (13 allocations: 1.38 KiB)
        New mixed:  3.791 μs (13 allocations: 1.38 KiB)
        Old homogeneous:  1.540 μs (2 allocations: 560 bytes)
        New homogeneous:  979.700 ns (2 allocations: 560 bytes)
```

The proposed method is uniformly faster for mixed tuples and for `NTuples` of small, moderate and large sizes. For tuples of moderate size (~10) the difference is very significant.

I agree that additional constant `N=20` is arbitrary. I have chosen it so that compile time is still low, but it can be changed if such a high `N` is an issue. I am not qualified to fully understand if having too many methods for such a function is problematic (as I believe that `reverse` for tuples is not a top priority functionality).